ヘッダーログイン判定の記述をlayoutコンポーネントに移動
